- updated deployment scripts - deployment script docs - OCD
