+++ scripts/lib/types.ts
@@ -1,0 +1,5 @@
+export type DeployResult = {
+    address: string,
+    abi: string,
+    opHash: string
+}
